% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACEPDATOS-package.R
\docType{package}
\name{ACEPDATOS-package}
\alias{ACEPDATOS}
\alias{ACEPDATOS-package}
\title{ACEPDATOS: Datos para analizar la conflictividad social}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Se trata de un paquete de datos para usar con la librería ACEP.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://agusnieto77.github.io/ACEPDATOS/}
}

}
\author{
\strong{Maintainer}: Nieto Agustín \email{agustin.nieto77@gmail.com} (\href{https://orcid.org/0000-0003-4467-873X}{ORCID})

}
\keyword{internal}
