% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_sismos_mdp.R
\docType{data}
\name{bd_sismos_mdp}
\alias{bd_sismos_mdp}
\title{Datos sobre conflictividad social en Mar del Plata (2011-2017)}
\format{
se trata de una tabla con 5132 observaciones y 3 variables:
\describe{
\item{id}{chr identificador de las acciones}
\item{fecha}{date años-mes-día}
\item{accion}{chr tipo de acciones}
}
}
\source{
\url{https://observatoriodeconflictividad.org/}
}
\usage{
bd_sismos_mdp
}
\description{
Datos sobre conflictividad social. Se trata de una serie de datos sobre conflictos sociales discriminados por tipo de acción, relevados por el Observatorio de Conflictividad de la UNMdP.
}
\keyword{datasets}
