% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflictos_con_paro_2006_2014_x_dependencia.R
\docType{data}
\name{conflictos_con_paro_2006_2014_x_dependencia}
\alias{conflictos_con_paro_2006_2014_x_dependencia}
\title{Datos sobre conflictividad laboral por dependencia estatal-administrativa}
\format{
se trata de una tabla con 27 observaciones y 5 variables:
\describe{
\item{dep}{chr dependencias administrativas: Estado Municipal, Estado Provincial, Estado Nacional}
\item{fecha}{chr a単os}
\item{conf_paro}{dbl frecuencia de conflictos con paro por a単o}
\item{huelguistas}{dbl frecuencia de huelguistas por a単o}
\item{jornadas_perdidas}{dbl frecuencia de jornadas perdidas por a単o}
}
}
\source{
\url{http://datar.info/dataset/conflictos-laborales}
}
\usage{
conflictos_con_paro_2006_2014_x_dependencia
}
\description{
Datos sobre conflictividad laboral. Se trata de una serie de datos sobre conflictos con paro discriminados por dependencia estatal-administrativa, relevados por la cartera de trabajo de Argentina.
}
\keyword{datasets}
