% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflictos_con_paro_2006_2014_x_reclamo.R
\docType{data}
\name{conflictos_con_paro_2006_2014_x_reclamo}
\alias{conflictos_con_paro_2006_2014_x_reclamo}
\title{Datos sobre conflictividad laboral por tipo de reclamo}
\format{
se trata de una tabla con 270 observaciones y 4 variables:
\describe{
\item{dep}{chr dependencias administrativas: Estado Municipal, Estado Provincial, Estado Nacional}
\item{ambito}{chr ámbito de los conflictos con paro: estatal, privado}
\item{fecha}{chr años}
\item{valores}{dbl frecuencia de conflictos con paro por año}
}
}
\source{
\url{http://datar.info/dataset/conflictos-laborales}
}
\usage{
conflictos_con_paro_2006_2014_x_reclamo
}
\description{
Datos sobre conflictividad laboral. Se trata de una serie de datos sobre conflictos con paro discriminados por tipo de reclamo, relevados por la cartera de trabajo de Argentina.
}
\keyword{datasets}
