% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pob_prov_censos_ar.R
\docType{data}
\name{pob_prov_censos_ar}
\alias{pob_prov_censos_ar}
\title{Datos censales sobre población}
\format{
se trata de una tabla con 24 observaciones y 3 variables:
\describe{
\item{prov}{chr provincia}
\item{pob}{dbl población}
\item{fecha}{chr año}
}
}
\source{
\url{https://www.indec.gob.ar/}
}
\usage{
pob_prov_censos_ar
}
\description{
Datos sobre población. Se trata de una serie de datos sobre población discriminados por provincia, relevados por el INDEC.
}
\keyword{datasets}
