% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflictos_con_paro_2006_2021_total.R
\docType{data}
\name{conflictos_con_paro_2006_2021_total}
\alias{conflictos_con_paro_2006_2021_total}
\title{Datos sobre conflictividad laboral total}
\format{
se trata de una tabla con 16 observaciones y 6 variables:
\describe{
\item{fecha}{dbl Años}
\item{conflictos_con_paro}{dbl frecuencia de conflictos con paro por año}
\item{huelguistas}{dbl frecuencia de huelguistas por año}
\item{jornadas_perdidas}{dbl frecuencia de jornadas perdidas por año}
\item{pro_huel_x_conf}{dbl relación huelguistas/huelgas}
\item{pro_jor_x_huel}{dbl relación jornadas perdidas/huelgustas}
}
}
\source{
\url{https://www.trabajo.gob.ar/downloads/estadisticas/bel/1989014.xlsx}
}
\usage{
conflictos_con_paro_2006_2021_total
}
\description{
Datos sobre conflictividad laboral. Se trata de una serie de datos sobre conflictos con paro, relevados por la cartera de trabajo de Argentina.
}
\keyword{datasets}
